Chat: moved the testHelpers module to the __tests__ subdir.
