+++ src/chat/__tests__/cards/NotesCard.test.tsx
@@ -2,11 +2,14 @@ import { waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
-import withProviders from '../../../components/withProviders';
+
 import { renderWithRedux } from '../../../utils/test-utils';
+
+import withProviders from '../../../components/withProviders';
 import NotesCard from '../../components/cards/NotesCard';
 import { initialEditMode } from '../../redux';
-import { currentUserMock, notesMock } from '../../testHelpers';
+
+import { currentUserMock, notesMock } from '../testHelpers';
 
 describe.skip('<NotesCard />', () => {
   const NotesCardWithUser = () => <NotesCard />;
