+++ src/chat/__tests__/cards/NotesInput.test.tsx
@@ -1,12 +1,15 @@
 import { fireEvent, waitFor } from '@testing-library/react';
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
-import withProviders from '../../../components/withProviders';
+
 import { renderWithRedux } from '../../../utils/test-utils';
+import { sendNote, editNote } from '../../../utils/api';
+
+import withProviders from '../../../components/withProviders';
 import NotesInput from '../../components/cards/NotesInput';
 import { initialEditMode } from '../../redux';
-import { currentUserMock, notesMock } from '../../testHelpers';
-import { sendNote, editNote } from '../../../utils/api';
+
+import { currentUserMock, notesMock } from '../testHelpers';
 
 jest.mock('../../../utils/api');
 const sendNoteMock = sendNote as jest.Mock;
