+++ src/chat/__tests__/cards/UserInformation.test.tsx
@@ -1,9 +1,12 @@
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
-import withProviders from '../../../components/withProviders';
+
 import { renderWithRedux } from '../../../utils/test-utils';
+
+import withProviders from '../../../components/withProviders';
 import UserInformation from '../../components/cards/UserInformation';
-import { chatMessagesMock, currentUserMock } from '../../testHelpers';
+
+import { chatMessagesMock, currentUserMock } from '../testHelpers';
 
 describe.skip('<UserInformation />', () => {
   const UserInformationWithUser = () => <UserInformation />;
